+++ internal/config/jevstyle_test.go
@@ -1,6 +1,11 @@
 package config
 
-import "testing"
+import (
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+)
 
 func TestJevstyleInUse(t *testing.T) {
 	t.Parallel()